[feat]네비게이션 바 수정
FixedComponet => 버튼누르면 아래 영역에 나오고 종료하면 사라지는 컴포넌트 추가
